Fix: Change phone number from int to long long to handle large numbers correctly

+++ src/header.h
@@ -30,7 +30,7 @@ struct Record {
     int userId;
     char name[100];
     char country[100];
-    int phone;
+    long long phone;
     char accountType[10];
     int accountNbr;
     double amount;
